264. API Authorization

+++ app/Http/Controllers/Api/V1/PostCommentController.php
@@ -57,6 +57,8 @@ class PostCommentController extends Controller
      */
     public function update(BlogPost $post, Comment $comment, StoreCommentRequest $request)
     {
+        $this->authorize($comment);
+
         $comment->content = $request->input('content');
         $comment->save();
 
@@ -70,6 +72,8 @@ class PostCommentController extends Controller
      */
     public function destroy(BlogPost $post, Comment $comment)
     {
+        $this->authorize($comment);
+
         $comment->delete();
 
         return response()->noContent();
